remove unused function createEmptyMMContainers
Summary: createEmptyMMContainers was used when serilizing for persistence for the reason of backward compatibility. A copy of empty MMContainers will be created and serialized as unevictable mm containers and then written to metadata so that previous CacheLib versions can restore from such a serialization. This function is no longer used so can be safely removed.

Reviewed By: haowu14

Differential Revision: D37027289

fbshipit-source-id: c46e7136aaedd8984a22a706dc368f68218d9e73