Clean up the long initializer list in CacheAllocator
Summary: Clean up the initialization list of the three original constructors of CacheAllocator. Combining the initialization logic into one delegated constructor with some helper functions to improve the readability. The initialization logic is now in one place which should be able to reduce the error where new member or changes to the member initialization is not synced in all the constructors.

Reviewed By: haowu14

Differential Revision: D37012499

fbshipit-source-id: 326528498307292d95a01145443216e3b56bec69