Ensure each clean and outstanding region has a in-mem buffer associated with it
Summary: We used to allocate in memory buffer statically as a list of buffers with a fixed size. The in memory buffer allocation is now changed to dynamic allocation where no predefined number of in memory buffer is needed from the user. Corresponding memory buffer will be allocated and released when attaching to or detaching from a region.

Reviewed By: haowu14

Differential Revision: D37043575

fbshipit-source-id: 61a10e9724929f822aa374f4464260719c381896