added support for multi-threading and trace amplifications for replay generator
Summary:
Since the workload traces are sampled at a very small rate (e.g., 1/50000), it is not possible to
get meaningful results from simply replaying the captured traces on real hardware. Also,
previously, the multi-threading was not supported for the replay generator, implying that the replay
generator is not adequate to emulate the races between concurrent accesses prevalent in real
production environment.

To solve those problems, this change implements the support for amplifying the traces along with
implementing the multi-threading support. Specifically, when the specified thread count is larger
than 1, the newly added generator, AmplifiedReplayGenerator, instantiates as many threads as
specified, where each thread opens the same trace file and replays each requests. In doing so, the
request key is apppended with the unique thread ID such that the keys generated by each thread are
never identical.

Reviewed By: therealgymmy

Differential Revision: D37569812

fbshipit-source-id: 85d1bae39c224039df87ec11dd13781397bb4a6a