Throw if there is a type mismatch in creating JSON config
Summary: In JSON config, array should be wrapped with brackets. We did not throw or print any error message for array type mis-match. For int, double, bool, and string, throw is implemented in asInt(), asDouble(), asBool(), and asString(). We throw the same TypeError exception for array when isArray() returns false.

Reviewed By: therealgymmy

Differential Revision: D38256432

fbshipit-source-id: e8f45ff4b0322fc4cc431712b71c1d3e7be23040