Adding EventBaseThread for Asynchronous Implementation
Summary:
We use EventBaseThread to implement the asynchronous request model for CacheBench. There will be one event base thread per worker to handle the asynchronous requests.

In this design, the worker thread is responsible for getting the requests from the workload generator and process them. If the requests need to be processed asynchronously, it will be passed to the event base thread.

Reviewed By: therealgymmy

Differential Revision: D38443609

fbshipit-source-id: 192cdb111c43b849937898c812a71cd107ee62e4