Adding Asynchronous Find Support in CacheBench's Cache Wrapper
Summary: Original CacheBench only support synchronous find where it will block and wait for the handle to be ready from CacheLib. In order to perform asynchronous find, we need to add another API in CacheBench. This API will return a SemiFuture of the handle which can be executed by an executor later.

Reviewed By: jaesoo-fb

Differential Revision: D38439440

fbshipit-source-id: b636d325beb6987f2db0ed66b603573d2e4611e4