do not persist/recover expired items
Summary:
# Context
Entrepot Cache service is suffering from sporadic OOMs during restarts. This is attributed to the fact that when we restore content of the cache, cache OOM protection is not enabled yet so depending on the state of task Mem usage in start up, we might end up OOM-ing.

To reduce the number of items being saved and restore we don't need to persist or restore it If item is already expired. This should help with OOM given that total number of items restored will be lower.

Reviewed By: therealgymmy

Differential Revision: D38018678

fbshipit-source-id: ee7aa20a7f4145992f6a38562c4e88ad8c3d7049