Return empty stats if allocation class is not initialized
Summary:
In getPoolStats, do not access nullptr and SIGSEGV if allocation class is not initialized. In this diff, we export an empty class stat.

This could happen if some thread collects stats before the cache is initialized or after shutdown. This might have caused a release blocker in MC when we migrate to fb303 counter, which gets collected in the background regardless of the cache's status. See the details in T130879960.

I think this function should be safe to call since it is a public function.

Reviewed By: therealgymmy

Differential Revision: D39190693

fbshipit-source-id: 1958771c03cf367e3448d5c85938f008e6bfe9b9