Add couldExistFast API
Summary:
This introduces the API as outlined in https://fb.workplace.com/groups/cachelibteam/posts/1336237183565284

A couldExistFast() API returns true if a key is potentially in cache. There is a non-zero chance the key does not exist in cache (e.g. hash collision in NvmCache). This check is meant to be synchronous and fast as we only check DRAM cache and in-memory index for NvmCache.

----
1. Add couldExist API
2. Add cachebench support for it and a config
 ---
What's not in this diff. Consistency check support. It seems a bit more complex than I initially anticipated to add consistency check. Basically we need to differentiate couldExist and get from the checker (TLDR: we need to differentiate couldExist -> couldExist and couldExist -> get).

Reviewed By: haowu14

Differential Revision: D37479631

fbshipit-source-id: 5471a7c6984142a624cbc208d4cbc86a30d40dcd