Refactor DynamicRandomAP ThrottleParams
Summary:
We have a struct called ThrottleParams, which contains all the parameters necessary for updating the parameters to perform throttling in DynamicRandomAP. This struct is not thread save, so we intend to copy it on read under a lock.
Inside this struct, there are two types of parameters:
1) The ones that needs to be used on every admission tests.
2) The ones that are used only when updating the params (once per minute).

In this diff, we separate these two types of params into two different structs. Only copy 1) on admission tests and use 2) only when update and exporting counters (WriteStats).

Reviewed By: therealgymmy

Differential Revision: D39356661

fbshipit-source-id: d99309380d0904f4b143bca18ba5a80bf7db44ee