ItemStatsExporter: API change for getSampleItem
Summary:
Previously, the cachelib was uploading CacheItem sampled only from DRAM cache to Scuba table, which
limits the usability of the feature. As an intermediate step to extend the feature to support NVM
cache, this commit changes the API to allow the sample can be returned from any memory location.
Previously, the returned sample is supposed to be in the Slab, implying the sample items from NVM
cache could pollute the DRAM cache unnecessarily.

Specifically, this commit introduce a new class named SampleItem which wraps the CacheItem and
changes getSampleItem() to return it instead of the item handle. For DRAM item, the item handle will
still be held by the IOBuf in the SampleItem so that the reference to the CacheItem can be dropped
appropriately when the client drops the reference to the SampleItem.

Reviewed By: therealgymmy

Differential Revision: D38957233

fbshipit-source-id: cc6f6c5225dcd60b2fc3794307d35220a59aeeff