Speed up CMS reset()
Summary:
There is a latency spike on hour boundary for bigcache origincache. We suspected it related to ML admission policy's access tracker because it is switching to the next hour. I found the root cause to be CMS reset().

To speedup the reset(), I evaluated the current impl and memset and picked this implementation. Details in the benchmark added.

Verbose debugging details:
https://docs.google.com/document/d/1so1B4af2IDZWQlTsYfUvLM_-NZWM3r7BuT2-QRYzZtM/edit?usp=sharing

Reviewed By: therealgymmy

Differential Revision: D39781283

fbshipit-source-id: f976b5a208f7e78490a94980a2bb3c324381e527