ItemStatsExporter: upload random item samples from NVMCache
Summary:
Previously, the cachelib was uploading CacheItem sampled only from DRAM cache to Scuba table, which
limits the usability of the feature. This change extends the feature to sample item from NVM cache
as well in proporation to their sizes; i.e., the probability to sample items from DRAM, BigHash, and
BlockCache are determined by their sizes. In doing so, we should consider only those space actually
used to store the data. To do so, we added an interface to allow NvmCache to query the usage size of
each NVM cache engines.

Reviewed By: therealgymmy

Differential Revision: D38764598

fbshipit-source-id: 8f46705d42bfca56ce803495d771c66f6c4469f7