Add opRateBurstSize for TokenBucket-based req rate shaper
Summary:
Previously, there was a problem setting the burstSize as the same value as the genRate, which makes
it hard to control the burstSize separately for the token bucket-based req rate shaper. This change
fixes this problem by adding a separate configuration parameter opRateBurstSize for the toket
bucket-based req rate shaper.

Reviewed By: therealgymmy

Differential Revision: D37898476

fbshipit-source-id: 9b83669fbdf8509200629f1752e5116c46e0dcfc