Add is_intern to feature map
Summary:
In BigCache, we receive a flag as part of the request from Proxygen indicating whether it is for user/intern traffic (D40047373). The admission policy consumes this flag and tailors the cutoff score for admission accordingly.

This flag does not exist in cache-bench because it relies on workload generation from trace instead of running Proxygen.

This diff adds is_intern to the standard fields that the simulator reads from the trace. PieceWiseReplayGenerator then appends it to the feature map that gets sent with the requests to the cachebench admission policy. Jim can now consume this in his diff in FBDep.cpp: D40183454.

Reviewed By: jcipar

Differential Revision: D40457114

fbshipit-source-id: 53e44de26c97d042e60b53ca43c74203409bf722