Introduce bypass logic
Summary:
Allow users to bypass certain type of traffic.
This is needed for ucache, where we don't want to throttle any memcache traffic.

(This is intended to be a simplified version of D38667151, where throttling is applied on both types with a priority.)
The logic of tracking admitted bytes and compute logical write amplification does not change.

Reviewed By: therealgymmy

Differential Revision: D39529363

fbshipit-source-id: 3a248c131d120b195d91829b2c9d6a6b1f9aeaf0