fix bug truncating the slack space of NvmItem for item sampling and destructor
Summary:
When truncateItemToOriginalAllocSizeInNvm option is set to false (default), a client can utilize the
slack space in the slab allocation for a CacheItem. For this reason, we store the whole slab
allocation as the blob in NVM Cache. However, when we provide the NvmItem for item sampling and
destructor callback, we were truncating the slack space, causing a segfault when the client accesses
the NvmItem memory beyond its original alloc size. This change fixes this bug.

Reviewed By: therealgymmy

Differential Revision: D40814431

fbshipit-source-id: cb563ea8c0e77aabb0b021240a9c2deba65623cf