remove unnecessary re-enqueue for startReclaim
Summary:
Previously, we were enqueuing the reclaim jobs twice; one for selecting victim (called 'reclaim')
and the other for actual reclaim job (called 'reclaim.evict'). This was needed to capture the victim
RegionId, which is unnecessary. This change fixes this by defining the invalid RegionId as the
generalized lambda capture.

Reviewed By: therealgymmy

Differential Revision: D40920993

fbshipit-source-id: c5edfa3d8bc818da6eba3a51cb46ef3a4d684478