Add TTL support for trace fetcher and cachebench replayer
Summary:
This change adds the TTL support for the trace fetcher and cachebench replayer. Specifically, with
this change, the memcache trace fetcher fetches and emits the TTL field as well. The new output
format is:

op_time,key,key_size,op,op_count,size,cache_hits,ttl

Also, this change adds the TTL feature to cachebench. The cachebench can now parse the ttl as the
optional field and apply it in handling set operation. This change also makes some refactoring
around the trace parse logic.

Also, this change implements the support for the op type for ReplayGenerator.

Reviewed By: jcipar

Differential Revision: D41035565

fbshipit-source-id: 2ed1ac783205bd7f045ce12fc16877147e53279b