Add itemValue to cachebench
Summary: This adds `itemValue`, a `uint64_t`, to the cache trace. This is used to store the user/intern flag. We gave it an intentionally vague name, because we were concerned with putting stuff about user/intern split into the open source code. Other users of this code can use itemValue for whatever they want.

Differential Revision: D41134915

fbshipit-source-id: aa26ce715b70a4c95bb0678443832273522b4fa4