Fix bug of losing items across persistence
Summary:
Found this when debugging a a flaky test. The issue is that in the persistor, we terminate the worker if `breakOut_` is set to true. The master worker sets this to true when all the items have been dequed from folly MPMC queue is empty. However, when MPMC queue is empty, it does not mean the workers have finished processing. Terminating early can cause us to lose items across persistence. Reading the original code, this should be a minor issue affecting only the last item in the queue.

The fix is to always check break out condition first before checking the queue. This way, the master thread and the worker threads can synchronize correctly and workers will only break out after it has dequeued everything and has finished processing the last item.

Reviewed By: jiayuebao

Differential Revision: D42461685

fbshipit-source-id: e542efe6f014ce89c7204e4e1854d84e9a1316bd