Fix flaky test - AddChainedItemMultiThreadWithMovingAndSync
Summary: Some times the moving attemps are exhausted and we ended up evicting items. Other times we get unlucky and release a slab with no items in it. The changes increase moving attempts to a million times per item, and also only checks for if items got evicted after the rebalancing is done. (Lack of moving could also mean no item was found in a rebalanced slab)

Reviewed By: haowu14

Differential Revision: D42514120

fbshipit-source-id: 52f8eb8b232ea2187cee118ae59af571f279761c