awake GetCtx waiters outside of FillMap lock
Summary:
According to the strobelight from cachebench run with hybrid cache configuration (T140823697), we
spend more than 9.5% of time on removeFromFillMap() code (at around 83K/s lookups from NVM), which
runs with holding the FillMap lock. However, this could make the critical section unnecessary long,
which could cause the high contention for the FillMap lock (sharded with 8K entries). This change
fixes this problem by releasing the lock early, and making the wakeups for the waiters without
holding the lock.

Reviewed By: therealgymmy

Differential Revision: D42212434

fbshipit-source-id: d843ac3a2a237968c9cfc9bf47ebd0475467820c