Don't use static to get cache size
Summary:
If there are multiple cachelib instances running on the same process, the variables will be shared and the ratio between DRAM and NVM could be more.
More specifically, if a DRAM-only and hybrid cachelib co-exists and the hybrid cache is initialized and triggered this first, the DRAM-only item sampler would run into nullptr because it would believe it has NVM cache while its a nullptr.

Reviewed By: jaesoo-fb

Differential Revision: D42690076

fbshipit-source-id: 9c65c73705c22903e19e8144cd41c76b18cfe423