ReplayGenerator: support for amplification factor in separation from threads count
Summary:
Currently, the amplification factor of ReplayGenerator is bound to the number of threads, since we
have implemented the amplification feature by appending thread ID to the key as the suffixes.
However, for better flexibility and usability, it is better to separate the amplification from the
number of stressor threads. This change decouples the amplification factor from the thread counts.
Specifically, one can specify the amplification factor "ampFactor" as part of
"replayGeneratorConfig".

Reviewed By: therealgymmy

Differential Revision: D42932839

fbshipit-source-id: 4f43ea936286e81e5eacbf378fd6b5a1ac60ec40