fix bug not terminating with numOps limit
Summary:
After D42932839 (https://github.com/facebook/CacheLib/commit/9796f9b7f56f1a2e2d77fdd7a6798470c0bdf277), each stressor thread is associated with dedicated request queue. However, when the
stressor is terminated after numOps limit, it was not notified to the replay generator thread,
causing it to spin forever to insert requests to the request queue without consumer. This change
fixes this bug by implementing the markFinish() which set the flag indicating the stressor is
terminated.

This change also make the appropriate names set for the stressors and the replay generator thread to
ease the debugging.

Reviewed By: therealgymmy

Differential Revision: D43163406

fbshipit-source-id: 328e54ba9da0aa3d35e0939fdd11e51a7abecb89