rename ReplayGenerator to KVReplayGenerator
Summary:
Current ReplayGenerator is specific to the trace in simple <key, value> format. Since we would want
to have more ReplayGenerator (e.g., TAOReplayGenerator, WSCacheReplayGenerator ) in the future, it
would be great to change the name to avoid any confusion in the future. To do so, this commit
changes the name of ReplayGenerator to KVReplayGenerator to be specific.

Reviewed By: therealgymmy

Differential Revision: D43170955

fbshipit-source-id: d5947ae87e839b5269b42b9200bc8352e17d8ddc