fix flaky tests in NvmCacheTests
Summary:
This change fixes following flaky tests in NvmCacheTests.
* NvmCacheTest.Delete
* NvmCacheTest.NvmEvicted
* NvmCacheTest.EvictToNvmGetCheckCtime

The root cause of the failures are essentially the same as D42443647 (https://github.com/facebook/CacheLib/commit/5e7ff9ab28cc00c74b19b176885d6af2e3d27d60) which fixed the problem for
NvmCacheTest.EvictToNvmGet; we are inserting enough items that could be spilled to NVM cache, where
the NvmCache::put() can be dropped and the item is evicted completely when the delete operations
(and tombstones) issued as part of the insertion are still outstanding. In order to fix the problem,
this change flushes the NVM cache periodically during the insertions.

Also, since this could cause more regions are used, the size of NVM cache needs to be increased.
This change bumps the default size of NVM cache to 200MB (previous 100MB). Also, the size of persist
storage used in the test PersistenceCache has been bumped by 100MB accordingly, i.e., from 400MB to
500MB.

This change addresses the github issue https://github.com/facebook/CacheLib/issues/169

Reviewed By: therealgymmy

Differential Revision: D43592888

fbshipit-source-id: f0968884eb39fb5728b59129e98345df3240f01e