Introduce 'markedForEviction' state for the Item. (#183)
Summary:
It is similar to 'moving' but requires ref count to be 0.

An item which is marked for eviction causes all incRef() calls to that item to fail.

This will be used to ensure that once item is selected for eviction, no one can interfere and prevent the eviction from suceeding.

'markedForEviction' relies on the same 'exlusive' bit as the 'moving' state. To distinguish between those two states, 'moving' add 1 to the refCount. This is hidden from the user, so getRefCount() will not return that extra ref.

Pull Request resolved: https://github.com/facebook/CacheLib/pull/183

Test Plan:
flatmemcache shadow result:

It seems there is a slight latency regression in allocation (~1 micro second)
https://fburl.com/ods/u3isn8b4
It's probably small enough to ignore. But we can come back if this starts showing up in large scale in production.

Imported from GitHub, without a `Test Plan:` line.

Reviewed By: therealgymmy

Differential Revision: D42089974

Pulled By: haowu14

fbshipit-source-id: 90477faa9443080a7ddefa37a524965a04b6f084