WorkloadGenerator: fixed a bug in scaling key popularity distribution
Summary:
There was a bug in scaling the number of keys in WorkloadGenerator. Particularly, even when user
wants to scale the number of keys larger than the sum of the number of keys in the popDistFile, the
generated key population is a lot less than specified, but only close to the sum in popDistFile.
This is due to the bug in FastDiscreteDistribution. Specifically, when it initializes the popularity
buckets and the weights, there was a bug not initializing the key offset of bucket correctly; it
was not considering the scaling factor into account. This change fixes the bug.

Reviewed By: therealgymmy

Differential Revision: D44144006

fbshipit-source-id: 1391b63b319598ac9d3e1307a86ad22b550bc2eb