rust: fix potential integer truncation for oversized cache allocations
Summary: Cachelib can only accept allocations up to 4MiB, and as such the size parameter is a `uint32_t`.  However, the Rust API currently accepts `usize`, which may be truncated when converted to `uint32_t`.  Make this conversion explicit and an error if the `usize` is larger than what will fit in a `uint32_t`.

Differential Revision: D44186153

fbshipit-source-id: f5a57e5447cf93211a259f8360e06cafa7564158