properly support a file-backed nvm cache
Summary:
Previously, in order to use a file as the backing store of nvm cache, user needs to create the file
with the given size a priori, hurting the usability of the feature. This change fixes it so that the
new file with the given size is automatically created if not exist. Also, if the file exists with
different size, this change will allow the cachebench to ftruncate or fallocate to the given size
automatically.

Reviewed By: therealgymmy

Differential Revision: D44154281

fbshipit-source-id: c622102d4d67b64da83dcad4fdade66473d70a1f