implemented ampFactor for PieceWiseReplayGenerator
Summary:
Currently, PieseWiseReplayGenerator does not support amplifying the trace in terms of key
popularity, limiting the applicability of cachebench simulation; e.g., use cachebench for SSD
qualification in terms of latency, WAF, write rate, etc. This change implements the feature to
amplify the trace key popularity using the same method we used for KVReplayGenerator (D37569812).
Specifically, this change adds a new replay generator option called 'ampFactor' and the
PieceWiseReplayGenerator replaces the last four '0's at the end of trace with the stream ID which is
in the range [0, ampFactor). This implementation assumes that the trace is anonymized using newly
added fetch_anonymize_trace in D44356554 which encodes the key as unique 11B string with appending
'0's upto the original key size.

Reviewed By: therealgymmy

Differential Revision: D44810169

fbshipit-source-id: 36b027d2884411d8964ba5a1064c79fae604498d