fix bug missing items while persisting ObjectCache
Summary:
There was a bug that some objects can be dropped while persisting cache. This is caused by misuse of
MPMCQueue API, where the return value of the non blocking write, i.e., MPMCQueue::write(...) was
ignored when the persister pushes the items to the work queue. The write can easily fail because the
MPMCQueue is a bound queue and we were creating the queue with the capacity of 1000. This bug was
not detected by the unit test because we were only inserting 800 items which is less than the
capacity of 1000 due to another bug T133925389 where only 1 shard is created if the item limits is
less than 13797.

This change fixes the bug by changing the non-blocking write call to the blocking call; i.e,
MPMCQueue::blockingWrite(...).

This change also fixes another problems in the unit test:

1. insertWithPoolId did not work as intended
We were inserting item directly to the specific pool using `CacheAllocator::allocate(...)` API.
However, when the items are restored, each item will be inserted into different pool determined by
the key hash, causing unexpected behavior.

2. The find call for randomly shuffling the items' order did not work
We were trying to shuffle the item orders in MMContainer by calling find() for randomly selected
objects. However, the key has different format than those used for insert, so the find() calls were
just no-ops.

Reviewed By: therealgymmy, jiayuebao

Differential Revision: D44753950

fbshipit-source-id: 29d260ba3d606e3258696888b6320f40f9f6e8c4