Support metadata piece
Summary:
[Piece-wise caching wiki](https://www.internalfb.com/intern/wiki/Traffic/CDN/CDN_Core/BigCache/#piecewise-caching).

{F909022315}

Currently, the simulator is not creating a metadata piece, only header + body pieces. This diff fixes that.

(The reason we're fixing this is to enable accurate **per object access history tracking** in the simulator. The tracker increments only on pieces with **key == base_key**, i.e. the metadata piece for large objects handled piece-wise, or the single piece for small objects that fit into a single piece).

Reviewed By: tang-jianfeng, jaesoo-fb

Differential Revision: D44990407

fbshipit-source-id: 5eafb033c5833d3b32b5b6880250cc86de94c9e8