fix build errors caused by deprecated Node.js versions for actions
Summary:
All of the build workflows are failing with below errors.

Node.js 12 actions are deprecated. Please update the following actions to use Node.js 16: actions/checkout@v2, actions/upload-artifact@v2. For more information see: https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/.

The Node.js version has been updated to 16 in v3 of the relevant actions, checkout and
upload-artifact. This change also make the workflow to use the v3 of actions.

Reviewed By: therealgymmy

Differential Revision: D45092631

fbshipit-source-id: 44377905e3041c0fd070d10ac093f23e8a1ce346