fix bugs in sizing the pools and the placeholders of ObjectCache
Summary:
This change fixes three bugs in sizing the objectcache pools with regards to the number of entries
limit.

1. Round error in computing cache size
The slab size is not multiple of the alloc size in most of the cases, so there is always a
fragmented space (16B, 4MB % 304 where 303 is the alloc size) in each slab which is ignored
in the following computation.
```
const size_t l1SizeRequired =
    util::getAlignedSize(config_.l1EntriesLimit * l1AllocSize, Slab::kSize);
```

We will see the off-by-one error accumulated at every multiples of 262144 or so;
`13797 * ceiling(304 / 16) + 1= 262144` where 13797 is the number of allocs per slab

2. l1NumShards configuration value is ignored when the l1EntriesLimit is small
The perPoolSize computation in below code is kSize at minimum, causing l1NumShards_ is rounded down.
```
  l1NumShards_ = std::min(config_.l1NumShards,
      this->l1Cache_->getCacheMemoryStats().ramCacheSize / perPoolSize);
```
So, l1NumShards_ could have different value if the l1EntriesLimit is less than or equal to
`13797 * (l1NumShards - 1)`.

3. imbalanced entries limits between pools when `l1NumShards` is not 1
When we compute the `perPoolSize`, we are not aligning it to the slab size
https://fburl.com/code/goaqv02k. This means that the cache size might not be integer multiples of
l1NumShards in terms of the slab counts. Since slab is the unit of allocation for memory allocator,
each pool could have different number of slabs, i.e., one slab more or less. Also, it is not
considered in computing the required number of the place holders for limiting l1EntriesLimit.

Reviewed By: therealgymmy, jiayuebao

Differential Revision: D44754364

fbshipit-source-id: ec28d623977e56e4e2d723edac0a3b2a41c35001