add ppa:git-core/ppa to install updated version of git
Summary:
Since we updated the github workflow's checkout action to v3 D45092631, the ubuntu-18 build was
failing at the checkout step with below errors.

```
The repository will be downloaded using the GitHub REST API
To create a local Git repository instead, add Git 2.18 or higher to the PATH
```

In order to fix this issue, this change adds the ppa:git-core/ppa so that the latest version (2.14
as of now) of git instead (default git version in ubuntu-18 is 2.17.1).

Reviewed By: therealgymmy

Differential Revision: D45242734

fbshipit-source-id: 382fadd56502aef9dccd56d1595b936f040178a0