fix segfault due to off-by-one error accumulation in FastDiscrete.h
Summary:
Previously, we were accumulating the off-by-one error in the next offset of a bucket in
transforming/scaling source buckets to target same-weight buckets for FastDiscreteDistribution.
Particularly, the overflow in which the total range of output distribution could exceed the
specified range could happens when there are some hot source bucket that spans multiple same-weight
buckets. This change fixes the bug by computing the next offset using the total number of objects
accumulated so far.

Reviewed By: therealgymmy

Differential Revision: D46091499

fbshipit-source-id: ee91e4b3c29ac49d67ba9f747f12c73fe9472afa