support for dynamic header format
Summary:
This change implements the feature for TraceFileStream to parse the trace line (i.e., setNextLine)
and allow the replay generator to retrieve the field with the field ID (i.e., getField) later on. In
order to support this, the replay generator provides the column names to field ID mapping table when
constructing the TraceFileStream.

This change also modifies the KVReplayGenerator to apply the new feature. With this change the
memcached traces can be shared in any format in forward compatible manner as long as the appropriate
header line is specified.

This change also adds an unit test to verify the new format and backward compatibility.

This change discards (most if not all) and reimplements D41403750.

Reviewed By: therealgymmy

Differential Revision: D46456693

fbshipit-source-id: d5b044816675127a740b20f62bba0563f3079916