Adds the support for Flexible Data Placement(FDP) over NVMe into the Cachelib.
Summary:
This commit allows the BlockCache and BigHash to segregate their data streams in physical NAND media
through the FDP placement Identifiers. With this changes, the Cachelib can reduce the Device Write
Amplification (WAF) in high SSD utilization scenarios such as "nvmCacheSizeMB" above 50% of the SSD capacity.

It introduces a 'placementHandle' concept, which can be used by both BC and BH of Navy on device write() calls.
The 'placementHandle' have to be allocated from the device. The placement capable devices will return a valid
handle, while the non-capable devices will return a default handle. 'placementhandle' is envisioned as an
abstract concept for all data placement technologies; not just for FDP.

This introduces a new class NvmeFDP in Device layer, since FDP semantics need to be handled differently from the
regular block interface. At the moment, NVMe directives can be passed to the kernel only through nvme-passthrough
interface. So this commit introduces two new classes - NvmeInterface and IOInterface.
NvmeInteface is an abstraction for all Nvme specific details, many of which can be eventually moved to some library.
And IOInterface is an abstraction around the IO interfaces supported - IOUring Passthrough is used in this version.

The user needs to configure an NVMe char device instead of the regular block device to enable the FDP mode.
(Ex:"nvmCachePaths": ["/dev/ng0n1"])

It does NOT incorporate the FDP Reclaim Unit(RU) boundary awareness during allocation and de-allocation of blocks.
That feature can be incorporated later if required.

Signed-off-by: Arun George <arun.george@samsung.com>