Update worker interval to 0 seconds when they're stopped
Summary:
Today sometimes customer will stop cachelib workers and not start new ones. In this case, we should just update the worker interval setting to 0 to reflect they're stopped.

This behavior us down the wrong path when investigating Tippy TTL behavior (we thought the reaper was running when it wasn't)

Reviewed By: jaesoo-fb

Differential Revision: D47033569

fbshipit-source-id: 7bf8597882f91b91b881d81465ba36a9f9b2628a