Revert D47372992: Multisect successfully blamed "D47372992: [cachelib] Make periodic worker start and stop util functions" for test or build failures
Summary:
This diff is reverting D47372992
D47372992: [cachelib] Make periodic worker start and stop util functions by jiayuebao has been identified to be causing the following test or build failures:

Tests affected:
- [scm/grepo/integration_tests:commit-journaling - test-commit-journaling.t (run-tests)](https://www.internalfb.com/intern/test/844424993772291/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2429888
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: crassirostris

Differential Revision: D47462316

fbshipit-source-id: 7a296e78e73c4d2abfd7e8cd738d932b98510d4a