Add estimateWriteSize to Engine, EnginePair and Driver
Summary:
Working doc: https://docs.google.com/document/d/1lDgmwmXf4G57uPXv1UXNnHELeObHEm6Cv-DI5_ZPAu8/edit

As title. Provide a way to estimate the write size of a potential flash admission.
In BigHash this is always the bucket size. In BlockCache, it's the amount of space the item would take in the region.

Reviewed By: therealgymmy

Differential Revision: D47247515

fbshipit-source-id: a51116ce2dedce4d9ae60ce4cf53d98b32dc4cfa