fix OSS build error regarding folly::IoUring
Summary:
OSS build is broken due to missing dependency on liburing and libaio since our AsyncDevice patch
D45932967 had been landed. In fact, io_uring is pretty new and thus only be supported on a few of
the most recent platforms (e.g., Ubuntu-22.04). Even in those platforms, the io_uring and liburing
might be outdated if folly::IoUring is dependent on the latest features like IORING_RECV_MULTISHOT.
To solve this problem, this change makes libaio installed in all of supported platforms and defaults
to libaio. The io_uring will only be enabled if it is installed manually on the platform.

Reviewed By: therealgymmy

Differential Revision: D48651333

fbshipit-source-id: dc7c39cb376089e86371ba8d22170170001df4ba