allow workflows can be triggered on demand by authorized users
Summary:
We are currently running workflows only on pull request and schedule-basis, making it take long to
validate any changes which is related to workflow builds. This changes adds "workflow_dispatch" for
the triggers of all workflows so that workflow can be dispatched on demand as well. Note that the
user must be a collaborator with write permission in the repository in order to trigger the
workflow with this option; see https://github.com/orgs/community/discussions/26053

Reviewed By: haowu14

Differential Revision: D48880563

fbshipit-source-id: faa61bf4f1d844625d793811788fd373e431358a