Optimize facebook::cachelib::MemoryPool::getStats
Summary:
**Created From Opportunity Instance: **[**529707119318263**](/overwatchx/opportunities/node?opportunity_instance_id=529707119318263)

**Background/help** for std::unordered_map: [https://www.internalfb.com/intern/servicedoctor/help/#std%3A%3Aunordered_map%2CCPU](/intern/servicedoctor/help/#std%3A%3Aunordered_map%2CCPU)

**BigGrep**: https://fburl.com/code/8u37tvnp

**Strobelight Icicle**: [https://fburl.com/scuba/strobelight_services/6q9mv034](https://fburl.com/scuba/strobelight_services/6q9mv034)

**Strobelight GraphProfiler**: [https://fburl.com/scuba/strobelight_services/qvs6ci1h](https://fburl.com/scuba/strobelight_services/qvs6ci1h)

**Strobelight Leaf File:Line**: [https://fburl.com/scuba/strobelight_services/47887xcv](https://fburl.com/scuba/strobelight_services/47887xcv)

**Call Graph**: https://fburl.com/cv/bf975z00

**Strobelight CPU**: [https://fburl.com/scuba/strobelight_services/fpmfbsq7](https://fburl.com/scuba/strobelight_services/fpmfbsq7)

**Rev b58d19e5ff**: [MemoryPool.cpp:514](https://fburl.com/code/7i1g13gz)

**Rev 01b7c0d046**: [MemoryPool.cpp:514](https://fburl.com/code/aq7rphkm)

**Rev 9e279593a8**: [MemoryPool.cpp:514](https://fburl.com/code/lraoazqi)

**Rev bf4d8c98d3**: [MemoryPool.cpp:514](https://fburl.com/code/z1knhdhz)

**Rev 5bc847b0ef**: [MemoryPool.cpp:514](https://fburl.com/code/ea909ifr)

As of 2023-09-10, the expected CPU win for optimizing this function is **0%**

_This Task was filed via _[_ServiceDoctor_](/intern/servicedoctor/)_ to help minimize work duplication via advertising perf investigations. As such, please keep the string __`"facebook::cachelib::MemoryPool::getStats"`__ in the contents of this Task so ServiceDoctor knows about it (even if this Task is closed)._

=====================[**INSTRUCTIONS FOR STATE CHANGE**]=====================

**Everything Has Done? Record Win To Mark It As Completed!**
You can navigate back to the [opportunity instance page](/overwatchx/opportunities/node?opportunity_instance_id=529707119318263) and record the win through the button in the action section.
Or you can directly close the task and the bot will reply on the task about how to record a win.
After you recorded the win, the task will be auto-closed and opportunity will be marked as COMPLETED.

**Invalid Opportunity? Blocklist It With Feedback!**
If you think this opportunity is not a valid opportunity, you could try to blocklist it through the action section in [opportunity instance page](/overwatchx/opportunities/node?opportunity_instance_id=529707119318263).
You will be asked to provide comments, blocklist category and a suspend time according to your assessment.
Your feedback will help us keep this community in a good shape.

**No Longer Available? Return This Opportunity To The Pool!**
If you think you no longer want to work on this task, you can change the owner to up to grab.
This will automatically return this opportunity to the pool for other people to take it over.

Reviewed By: jaesoo-fb

Differential Revision: D49175335

fbshipit-source-id: efa7f77b3fcc8b7aa20e5af182725f04c6e85fa8