remove test build options from all targets except ubuntu-22
Summary:
It seems that many of the targets including Centos-8.1/8.5 and Debian-10 are failing the builds due
to the no space left error on the container (e.g., https://fburl.com/0qlhm04z). This change attempts
to fix those build errors by omitting the test build option (-t) to see if that helps.

Reviewed By: therealgymmy

Differential Revision: D50246495

fbshipit-source-id: 326c9f265c170ff8752e9b199bfdf325ec9b98f9