use onFiber to determine whether to use async completion
Summary:
Previously, Navy Device determines which async IO completion mechanism (i.e., async vs sync wait) to
use base on the existence of eventbase on the issueing thread. This could cause a indefinite
blocking if the Device IO is executed on the main thread. For example, memcache main thread has
eventbase, but issues the cache shutdown path on the main thread synchronously, meaning the
eventbase cannot be processed until the shutdown path returns. In this case, the device IO issued to
flush the NVM cache cannot complete forever. So, instead of the eventbase, it is safe to rely on
onFiber() when determining which IO completion to use for given thread.

Reviewed By: therealgymmy

Differential Revision: D50203887

fbshipit-source-id: 22a89b454c54d555e0a504bbeae513fda1510ca4